requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
Django==4.0.3
django-rest-knox==4.2.0
djangorestframework==3.13.1
geographiclib==1.52
geopy==2.2.0
idna==3.3
pycparser==2.21
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
